Data.CSV.Conduit.Types

Documentation

data CSVSettings

Settings for a CSV file. This library is intended to be flexible and offer a way to process the majority of text data files out there.

Constructors

CSVSettings

Fields

csvSep :: !Char

Separator character to be used in between fields

csvQuoteChar :: !(Maybe Char)

Quote character that may sometimes be present around fields. If Nothing is given, the library will never expect quotation even if it is present.

defCSVSettings :: CSVSettings

Default settings for a CSV file.

csvSep = ','
csvQuoteChar = Just '"'

type Row a = [a]

A Row is just a list of fields

type MapRow a = Map a a

A MapRow is a dictionary based on Map where column names are keys and row's individual cell values are the values of the Map.